# [Perf] Small optimizations for silu_mul_fp8_quant_deep_gemm