## Purpose

Reorder some stride calculation and operations, rename to more descriptive variables, and decrease num_stages to 4 to help with expert token imbalance. Not a big deal but should help a bit.

```
python benchmarks/kernels/benchmark_silu_mul_fp8_quant.py

# main
GPU: NVIDIA H100 80GB HBM3
Config               Time(ms)   GFLOPS     GB/s      
--------------------------------------------------
E=  8,T=  32,H=1024    0.027     54.3     34.1
E= 16,T=  64,H=2048    0.027    457.2    287.5
E= 32,T= 128,H=4096    0.049   2188.1   1376.1
E=256,T=  16,H=7168    0.048   3534.2   2222.7
E=256,T=  32,H=7168    0.087   3978.8   2502.3
E=256,T=  64,H=7168    0.171   4064.7   2556.3
E=256,T= 128,H=7168    0.340   4157.8   2614.9
E=256,T= 256,H=7168    0.677   4244.9   2669.6
E=256,T= 512,H=7168    1.320   4291.0   2698.7
E=256,T=1024,H=7168    2.733   4185.6   2632.3

# this pr
GPU: NVIDIA H100 80GB HBM3
Config               Time(ms)   GFLOPS     GB/s      
--------------------------------------------------
E=  8,T=  32,H=1024    0.027     54.7     34.4
E= 16,T=  64,H=2048    0.028    455.7    286.6
E= 32,T= 128,H=4096    0.049   2179.3   1370.6
E=256,T=  16,H=7168    0.045   3788.7   2382.7
E=256,T=  32,H=7168    0.084   4130.8   2597.9
E=256,T=  64,H=7168    0.168   4149.8   2609.8
E=256,T= 128,H=7168    0.335   4229.8   2660.2
E=256,T= 256,H=7168    0.671   4282.4   2693.2
E=256,T= 512,H=7168    1.310   4325.2   2720.1
E=256,T=1024,H=7168    2.694   4245.8   2670.2
```

Eval shows the same score

```
# main
VLLM_ALL2ALL_BACKEND="deepep_low_latency" VLLM_USE_DEEP_GEMM=1 g2 lm_eval --model vllm --model_args pretrained=Qwen/Qwen3-30B-A3B-Instruct-2507-FP8,data_parallel_size=2,enable_expert_parallel=True --trust_remote_code --tasks gsm8k --num_fewshot 5 --batch_size auto
vllm (pretrained=Qwen/Qwen3-30B-A3B-Instruct-2507-FP8,data_parallel_size=2,enable_expert_parallel=True,trust_remote_code=True), gen_kwargs: (None), limit: None, num_fewshot: 5, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.8779|±  |0.0090|
|     |       |strict-match    |     5|exact_match|↑  |0.8719|±  |0.0092|

# tuning_silu_mul_fp8_quant_deep_gemm
VLLM_ALL2ALL_BACKEND="deepep_low_latency" VLLM_USE_DEEP_GEMM=1 g2 lm_eval --model vllm --model_args pretrained=Qwen/Qwen3-30B-A3B-Instruct-2507-FP8,data_parallel_size=2,enable_expert_parallel=True --trust_remote_code --tasks gsm8k --num_fewshot 5 --batch_size auto
vllm (pretrained=Qwen/Qwen3-30B-A3B-Instruct-2507-FP8,data_parallel_size=2,enable_expert_parallel=True,trust_remote_code=True), gen_kwargs: (None), limit: None, num_fewshot: 5, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.8779|±  |0.0090|
|     |       |strict-match    |     5|exact_match|↑  |0.8719|±  |0.0092|
```

## Test Plan

## Test Result

## (Optional) Documentation Update

---
<details>
<summary> Essential Elements of an Effective PR Description Checklist </summary>

- [ ] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [ ] The test plan, such as providing test command.
- [ ] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.
</details>